Remove old position retriever and rename to positions_retriever.
