+++ tests/event_finder/event_finder_rhodesmill/support/test_satellite_positions_with_respect_to_facility_retriever_rhodesmill.py
@@ -10,7 +10,7 @@ from satellite_determination.custom_dataclasses.satellite.international_designat
 from satellite_determination.custom_dataclasses.satellite.mean_motion import MeanMotion
 from satellite_determination.custom_dataclasses.satellite.satellite import Satellite
 from satellite_determination.custom_dataclasses.satellite.tle_information import TleInformation
-from satellite_determination.event_finder.event_finder_rhodesmill.support.satellite_position_with_respect_to_facility_retriever.satellite_positions_with_respect_to_facility_retriever_rhodesmill import \
+from satellite_determination.event_finder.event_finder_rhodesmill.support.satellite_positions_with_respect_to_facility_retriever.satellite_positions_with_respect_to_facility_retriever_rhodesmill import \
     SatellitePositionsWithRespectToFacilityRetrieverRhodesmill
